requirements.txt
asgiref==3.8.1
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
django-markdown==0.8.4
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
iniconfig==2.0.0
Markdown==3.7
msgpack==1.1.0
packaging==24.1
paystack-api==0.1.2
paystackapi==2.1.3
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.3
python-decouple==3.8
redis==5.2.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
